Fix assertions for Profman BootImageProfile tests
As part of the work done in aosp/2132472, we discovered that the
two tests in Profman for boot image profiles have different
behaviours (the first test creates a boot profile for a boot image,
while the second does not). We also realised that there are no
assertions around the call to ProfileCompilationInfo#Load, which is
expected to return true is the profile is loaded correctly and false
otherwise. This CL solves both issues, and makes the two test cases
more similar between themselves (maintaining the crucial difference
that the first tests has a single profile while the second has two
different profiles that need to be merged).

Test: atest art_profman_tests
Change-Id: I00620e97318540532971d8d0dd8d140576225a03